feat: Add newPet state and pass user ID
Moved newPet state to FullScreenModal to be closer to where it is
needed. User ID is required in the database to link the new pet profile
with the owner ID (user ID).

Additionally, style for submission button was updated and use button
instead of TextButton since Submit is a primary action. More info on
material.io.